backend/requirements.txt
django
djangorestframework
psycopg2-binary
requests
langchain-google-genai
langchain